Update to cocoapods 1.6.2 (#1696)
* Update to pod 1.6.2

* Fix lint error

+++ AirshipKit/AirshipKit/common/UAConfig.m
@@ -123,7 +123,7 @@ NSString *const UACloudSiteKeyName = @"site";
             _inProduction,
             self.appKey,
             self.appSecret,
-            self.logLevel,
+            (long)self.logLevel,
             self.defaultAppKey,
             self.defaultAppSecret,
             self.developmentAppKey,
